pbl6/requirements2.txt
absl-py==1.3.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
djangorestframework==3.14.0
drf-yasg==1.21.4
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.24.0
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==22.0
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tensorboard==2.13.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
